Rewrite order-updated to update sheet differently.
With this change, the spreadsheet reflects the net status of orders
after returns.  Canceled orders don't appear; partial returns result in
adjusted price and amount.

Lines are recognized by Stripe order ID and SKU, and are recognized
anywhere in the sheet.  New lines are added at the bottom.

This approach needs no locking to avoid conflict with office updates.